Feature: legislative news aggregation — RSS feeds, homepage widget, full news page
- Add 10 Louisiana RSS news sources with relevance scoring (0-20), industry tagging
  (12 categories), bill number extraction, and breaking news detection
- Sync script upserts articles every 15 min via Vercel cron; prunes articles >14 days
- NewsWidget on homepage: live indicator, breaking banner, 5 article rows with source badges
- Full /news page: tabs, industry/source filters, related bill pills, save preferences
- User preferences API (GET/PATCH) for personalized industry feed sorting
- News API routes: /api/news, /api/news/sources, /api/news/bill/[bill_number]
- Add "News" link to main navigation
- XML sanitization fallback for malformed RSS feeds (WDSU, LaPolitics Weekly)
- Apply 4 generate-summaries controls: haiku model, 3-paragraph prose format,
  full/partial quality filter, expanded banned phrases with rejection logging
- Fix search_bills() return type INTEGER → BIGINT in migration

+++ scripts/migrations/001-full-text-search.sql
@@ -78,7 +78,7 @@ CREATE OR REPLACE FUNCTION search_bills(
   page_size         INTEGER DEFAULT 25
 )
 RETURNS TABLE (
-  id                INTEGER,
+  id                BIGINT,
   bill_number       TEXT,
   title             TEXT,
   status            TEXT,
